requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.28

# Technical analysis (TA-Lib is optional - install separately if needed for better performance)
# ta-lib>=0.4.28  # Uncomment and install separately if desired
# If TA-Lib is not available, fallback implementations will be used

# Interactive Brokers (optional - only needed for live trading)
ib-insync>=0.9.86

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Configuration
pyyaml>=6.0
pydantic>=2.0.0

# CLI
click>=8.1.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3

